+++ packages/stylelint-config-itgalaxy/lib/config/rules/block.js
@@ -3,6 +3,7 @@
 'use strict';
 
 module.exports = {
+    'block-closing-brace-empty-line-before': 'never',
     'block-closing-brace-newline-after': 'always',
     'block-closing-brace-newline-before': 'always',
     'block-closing-brace-space-after': null,