+++ packages/stylelint-config-itgalaxy/lib/config/rules/declaration.js
@@ -8,6 +8,17 @@ module.exports = {
     'declaration-colon-newline-after': 'always-multi-line',
     'declaration-colon-space-after': 'always-single-line',
     'declaration-colon-space-before': 'never',
+    'declaration-empty-line-before': [
+        'never',
+        {
+            'except': [
+                'after-comment'
+            ],
+            'ignore': [
+                'after-comment'
+            ]
+        }
+    ],
     'declaration-no-important': null,
     'declaration-property-unit-blacklist': null,
     'declaration-property-unit-whitelist': null,